suppress "file changed" warning
This is causing unwanted email message from being sent when xivo-backup
is run from logrotate (which is run by cron). We still want more serious
warning/error message to be displayed.

One scenario when this happens is on the "/var/lib/asterisk" directory,
when you have a XiVO with lots of SIP phones and a low SIP default
registry expiration.