backup the /var/lib/consul directory
Previously, the /var/lib/consul directory was not backuped, but some
files in the directory were. With this commit, the backuped files are
still the same, but the directory is also backuped, so that if you do an
"rm -r /var/lib/consul" and then restore your backup, the owner /
permission on the /var/lib/consul directory are fine.